refactor data/get.ts
