+++ package.json
@@ -7,7 +7,7 @@
     "start": "next start",
     "lint": "next lint",
     "prepare": "husky install",
-    "prebuild": "ts-node data/get.ts"
+    "prebuild": "ts-node data/index.ts"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.4.2",
